// src/lib/types/wms-api/createAppointmentDto.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * ElasticWMS API
 * Warehouse Management System API - Web, RF, and Integration endpoints
 * OpenAPI spec version: 1.0.0
 */
import type { CreateAppointmentDtoAppointmentType } from './createAppointmentDtoAppointmentType.js';

export interface CreateAppointmentDto {
  appointmentNumber: string;
  appointmentType: CreateAppointmentDtoAppointmentType;
  carrierCode?: string;
  carrierName?: string;
  dockId: string;
  driverName?: string;
  driverPhone?: string;
  facilityId: string;
  notes?: string;
  referenceNumber?: string;
  referenceType?: string;
  scheduledEnd: string;
  scheduledStart: string;
  trailerId?: string;
  vehiclePlate?: string;
}