// src/lib/types/wms-api/dockAppointmentWebControllerListParams.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * ElasticWMS API
 * Warehouse Management System API - Web, RF, and Integration endpoints
 * OpenAPI spec version: 1.0.0
 */

export type DockAppointmentWebControllerListParams = {
facilityId: string;
dockId?: string;
date?: string;
status?: string;
};